+++ src/UI/UI.h
@@ -3,5 +3,6 @@
 #include "Core/Button.h"
 #include "Core/Menu.h"
 #include "Core/Pointer.h"
+#include "Core/TextArea.h"
 #include "Core/TextField.h"
 #include "Core/UIContainer.h"